Add profiles for relforge and enwiki
Manually tuned settings for enwiki and a set of queries identified as
problematic with all field.
Prod settings for boost templates are really not suited for these
settings and are way too high. I lowered their values onwiki but I
need to find a way to customize boost templates with config vars
otherwize this is not really testable with runSearch.php

Change-Id: Ib3b9320b79a16bf149e331ffa7be0aa8a36c4492