Simplify incoming_links counting from es query to mysql
Started working up this patch to allow changing the analysis chain used
for incoming_links counting. After more consideration will probably
add a new field to incoming_links rather than changing the existing
chain, but this patch still seemed usefull as it simplifys how the
incoming links are calculated and uses the canonical data source.

This might be slightly slower, but likely not much. The previous method
used an elasticsearch multi-search to do many documents at once. We
rarely index multiple documents at once though, only during a full
reindex that forces parsing (as opposed to in-place reindex). This
should be minimal load on the mysql slaves, the query is full covered
by the existing backlinks index. On a manual test of a few larger
pages the query to count ~150k incoming links to a page plus it's
50 reidrects is in the 100-150ms range.

Change-Id: Id7aaea59dd1ad5a7cb770c226cbf6ef7194233e8