Adjust brower test settings and add tie_breaker support.
Current settings seem to be too fragile and can cause cindy to
fail unexpectedly.
This patch tries to reduce the chances for this to happen by
lowering the norms impact and adjusting field boosts.

Added the possibility to customize dismax queries by
setting a tie_breaker. Avg field length is extremely
low for opening_text that we would need to either:
- completely disable norms on all fields
- set an insanely high boost so it can win in a dismax
- move it outside the dismax

Tie breaker seems to be a more reasonable approach and
could be usesul for production settings.

Change-Id: I2d12b3103afb2c08d470563c766d09105f6b0905