Use top_terms_blended_freqs with FullTextSimpleMatchQueryBuilder
Wildcards and the rewrite method used today are not designed to rank
results with weighted sum.
top_terms_boost generates a constant score query and for an unknwown
reason uses the queryNorm factor (issued from ClassicSimilarity?).
queryNorm is a constant query factor that was made in an attempt to
make lucene scores comparable, this factor is removed in lucene 7 and
not used by BM25.
queryNorm was so low that the generated scores were extremely hard to
use with a weighted sum.
Using top_terms_blended_freqs seems to force the use of the BM25
similarity, queryNorms are no longer used and scores are more
consistent. It permits to use the same rescore chain with and without
wildcards.

Bug: T143537
Change-Id: Idbce2e9646eb52fb707a6c86e6f4a40f8fd4e045