Ensure timeouts are always used
Works through the various code paths and ensures we are pretty
much always setting the client side timeout, and that we are
using the server side timeout where available. We should note though:

    The collector expires after the timeout and so it returns partial
    results. This doesn't necessarily stop parts of the query running on
    the node, eg a heavy rewrite on a wildcard query. The only way to
    interrupt things like this is to add checks, which would also make
    them all slow.

It seems this may or may not be as useful as I had expected, probably
still a good thing to do though.

Bug: T143006
Change-Id: I4f84dca4de56e939ff037e897f9de6f800e30976