Saneitizer: Check if revisions are up to date
Currently the saneitizer only checks if the page exists in the index,
and is in the correct index. Add a check that also makes sure we have
the latest version of the page in the index.

We have some bug reports, like T127788, that suggest we might
occasionally have out of date revisions in the search index.

Bug: T127788
Change-Id: If10fc6da10d51e2a12926a615ff10f1887bc2574