Add support for FLAG_SOURCE_DATA and defaultsort in completion suggester
This is still unclear if it will provide interesting results. It can
be useful in cases where the title does not start with a
representative word, but also to help mathing Surname, Firstname.
This should be carefully tested before activating it, the risk is to
hide other suggestions because the default suggestion will benefit
from the title score.

Change-Id: I1484c2e62788bedb57a42869a5fb25cd8f64482f
Depends-On: Ia41f8eeb9dd4f764543bdd4d71b7a50de8101101
Bug: T134978