Don't fail the whole reindexer thread when a single doc fails.
A single doc failure could cause the whole thread to fail, it can
happen in cases where some docs contain invalid coordinates that are
no longer accepted. We should not stop the whole thread.
This patch adds support for bulk response exceptions and will retry
only the docs that failed.
Since it won't fail in case of failure we try to detect earlier if the
resulting index can match the acceptable count deviation. This is to
avoid flooding elastic with unecessary index request when we know we
won't use the resulting index.

Change-Id: I6e811cefe78a1cb7988fbfb6f9049134c14f084b