Add script to compare cluster document existence
Not sure this is the best place for this script, it's somewhat
wikimedia specific although not overly so. It uses mget requests
against 2+ clusters to compare document existence. It spits
out saneitize.php calls necessary to bring the clusters in line
with eachother. It could be run from a shell something like:

  for wiki in $(expanddblist all); do
    python compare-clusters.py > fixes_required
    if [ -s fixes_required ]; then
      . fixes_required
    fi
  done

This was done in python, rather than php like much of this repository,
to take advantage of the multiprocessing lib which allows relatively
easily parallelizing the comparrison.

Change-Id: I2fa4f6e801ad2b0f342693430d92a5c2d3ff8a88