Rename 'detectors' to 'handlers' in super detect noop
For the super_detect_noop script to noop document updates within
20% of the original value it needs to read the configuration for the
handlers. This key was renamed in the search-extra plugin in I3e032ca3
but the change never made it's way into CirrusSearch. As such this has
basically been doing nothing for the past year and inducing more
indexing load than necessary.

Rename the 'detectors' key to 'handlers' so the plugin starts doing it's
thing again. This should, in theory, have some positive impact on the
indexing load on the search cluster.

Change-Id: I6a3be5d898d8b7e96086f0b5582b664faea5d2d5