Break words on semicolon for source_text.plain
Added a new word_breaker_helper char_filter specific
to the source_text field to avoid conflicts with normal
searches.
Created a new field class to crate a custom mapping for
the source_text field. Mostly to save space, the root field
does not seem to be used so there's no need to index it.

Bug: T145023
Change-Id: I49739a9ea3b0794303cfecf8edeb960fc786387a