Add support for ICU folding
The strategy is to update the generated config by switching
asciifolding to icu_folding and by forcing icu_folding (preserve
original) on plain analyzers.
This definitely makes the analysis config generation way harder to
understand.
Unfortunately the other approach would be to customize icu_folding on
a per language basis making it hard to have a config flag: if the
specified language is not customized then enabling icu_folding would
have no effect.
I hope this is just transitional, in the end I'd like to get rid
of asciifolding and maybe make the icu plugin a mandatory dependency.

Fix some oddities in the Suggest analysis config :
- only add an icu dependent filter if icu is available
- prefer setting analysis components before building the array instead
  of munging the whole config

This will cause the recycling process to create new indices.

NOTE: the patch is not really finished, it lacks proper per language
customization to filter out the characters that should be excluded
from icu_folding. I hope that this can be addressed gradually in
followup patches with feedback from native speakers.

Bug: T137830
Bug: T150799
Change-Id: I7fe1907147fe43406025bda331397d58c995158a