Fix cindy failures
Increased min word length for morelike querie to 2 to avoid matching
pages on the word "1" and make sure that ChangeMe is on the first
page.
Removed inc links boosting on one test to make sure we test text features
only.
Wrapped wfGetDB() in Checker::getDB() to force proper annotations with
IDatabase to resolve an annotation problem on Database or phan that is
not able to infer param types from interface.

Change-Id: Ifde771eb17ebeb7470964db617514b6cac1c89a1