Add timing marks to completion suggester
There is an unexplained spike in the completion suggester load times
with the wmf20 deployment and we don't really know what it is. Add
some timing marks and log them into CSRQ so we can hopefully at least
narrow down what code paths might be taking the increased time.

The names used here are pretty poor...it's too early for naming things.

Change-Id: I4360100dfe408833aaa48b81bf8ce0dfeb9501d9