Count the total hits, including duplicates, in completion suggester
The completion suggester was de-duplicating results before reporting
the total number of hits. To get a better idea of how many hits are
actually being returned instead report the number from before being
de-duplicated.

Was looking at this while trying to figure out how the loop that does
decode/dedup/scoring takes so much time, and it seems the actual number
of results processed rather than the post-dedup value would be more
meaningful to log. But perhaps there would be a reason to log both
the pre and post de-dup numbers?

Change-Id: I41c8e707cc44eebfeae8f31fb2ec9e85e7b46900