Move cluster override out of more like code
Prep work for moving more like into a keyword feature. More like needs
the ability to override the cluster that will be used (for performance
reasons).

This makes the overrides a bit more generic, and makes it obvious they
only apply to the full text search. I'm not sure there will be much use
for this outside of the more like functionality but never know.

Change-Id: Ic9a64ada2cd476fd5978fd89f57d8f34149bb236