Switch query parsing tests from Searcher to CirrusSearch class
To generate and test the appropriate more like query parsing fixtures we
need to start at CirrusSearch class, where the branching for more like
happens. I wasn't expecting it, but this also caused minor changes in
all the fixtures:

* Queries get appropriately trimmed of \n. Almost all the fixture files
  finish with a \n and we never bothered trimming it. This is probably
  a desired change
* file_text query portion removed from most queries. This also more
  accurately matches how real queries are built. For the most part only
  insource: queries have kept their file_text portion of the query.

There are so many files changed checking what changed is a bit of a
pain, but 90%+ of the changes are the exact same change. This can be
seen by using `git log -1 -p` and setting the console height to the size
of that change, then paging down changes only a few characters and it's
quite obvious the change is consistent (and the exceptions are obvious).

Change-Id: I977b990573fcb3ee6d04f37fca8f6ec41a4a7fba