Use standard query cache for interwiki, and add interwiki parsing tests
Adjusts how cirrusDumpQuery / cirrusDumpResult works to also handle
interwiki searching. With this new functionality add a query parsing
configuration profile for interwiki search, which will help ensure
upcoming refactors to use a MultiSearch for interwiki don't change
any behaviour.

Additionally switches interwiki search to using the default query result
caching mechanism. This avoids having to special case more code to
handle if interwiki results should be cached when dumpQuery / dumpResult
are used. This will be slightly slower, because the queries have to be
built before the cache key can hashed, but I doubt the hit rate is
particularly high on these anyways. As a side effect of this change we
will also be able to monitor the hit rate and make more informed
decisions about what to do with interwiki result caching.

Change-Id: I1eeaba9c4787a2ecb3e8277b81016497d37c1467