Add BM25 profiles to cirrus
These are the profiles used during the BM25 A/B test on english
wikipedia.
I was not sure where to put them but since there are rather large
code blocks it might be preferable not to bloat wmf-config.

This way in wmf-config we could only declare :
- weigths
- default rescore profile to use (inclinks or inclinks + pv)
- default query builder profile

Note the default weight values are doubled because we now properly
use BM25 as "the" default similarity which will inhibits the 0.5
coord factor on text features.
Phrase rescore field weights are doubled compared to the values used
in the test:

Assuming that:
  similiarty*0.5 + weight*phrase_rescore + weight*inclinks
is equivalent in term of ranking by applying a global factor 2:
  similarity + 2*weight*phrase_rescore + 2*weight*inclinks

Change-Id: I6476b67056de603effcd0669abf92b0cab724da6