Add some options useful for development
On relforge I always need to keep some patches to workaround the fact
we don't have access to mysql.
This patch allows to:
- ignore missing revision
- enable morelike queries by collecting the title from elastic

It'd save some time when rebasing on relforge.
I'm OK not to merge and keep it gerrit if we don't want to pollute cirrus
with such options.

Change-Id: I9f2c28ffba19dc5e4d596d73c5cbe21c4de10495