Update phan to use a config file
The latest versions of phan support configuration via a config file that
provides the files/directories to inspect, along with various settings
that effect what issues are raised. This also seems to now be detecting
a few more type mis-matches. This is prep work for moving the running of
phan into CI.

Change-Id: I6a0a949d5238a2d539d689336a61b4c65afb317b