Add configuration value to run interwiki load test
Before we roll out interwiki search we need to know that that impact on
our search cluster will be within acceptable limits. It is possible
there will be strong negative side effects, like pushing hot data out of
memory that now needs to be pulled from disk.

Patch creates a new configuration value that can be set to perform
interwiki queries, but not render their results, for a specified
percentage of full-text search requests to Special:Search. We can slowly
ramp this value from something small to 100% while monitoring the
cluster to safely test our ability to handle the increased load.

A matching patch will be required for the mediawiki-config repository to
both configure wgCirrusSearchInterwikiSources for all wikipedias, along
with settings wgCirrusSearchInterwikiLoadTest = 0.01, or some other
starting value.

The only thing not handled here is that some of our mocks rewrite the
search such that a search against content namespaces on wikipedia
performs an additional NS_FILE search against commons. Implementing that
feature is outside the scope of this patch, and will add some level of
uncertainty to interpreting the results of this load test.

Bug: T149740
Change-Id: Idb297c66dd626766a2ec6cb2b75d3c5a2a256afe