Do a better job normalizing stored query fixtures
While working on another patch that makes changes to some of the
queries, i noticed that the fixtures that are stored are really not
repeatably generated. Minor changes, that are not in any way important,
like the timestamp attached to 'now' or the ordering of some keys, don't
fail the test but when regenerating the fixture make for dirty diffs.

Attempt to do a better job normalizing the query before it gets written
out, such that minor changes to ordering don't make dirty diffs when
regenerating.

Also removes file_00{1..8}.expected. These files are no longer used, and
were recreated as file_00{1..8}.default.expected in a previous patch.

Change-Id: Id0911969d90c18b2fbe6dc41598d8930d50a3617