Make SearchConfig non-optional in the Searcher constructor
In a followup patch I needed access to the configuration in
InterwikiSearcher::__construct before calling parent::__construct. Could
have added the SearchConfig to only InterwikiSearcher, but it seemed
like things might be a little easier to follow if we always passed the
relevant configuration from the edges of the extension into the center
anyways.

Change-Id: I7ed08f4e00b78765d94d85d52fc2a187ffe7bbbe