Don't allow creating metastore from saneitizeJobs.php
It seems it's possible for the cluster to be in an odd state, where one
node has become partitioned off from the rest of the cluster, and that
node will tell our code that the index doesn't exist even when it does.
This code creates the index, and then the node has trouble re-joining
the cluster.

To be frank, this doesn't seem like a CirrusSearch problem and instead
is some sort of problem with elasticsearch itself. There should only be
one master, and creating an index should require an 'OK' from the
master. We've seen it happen though, so work around the issue by bailing
if the index doesn't exist, expecting it to always be created explicitly
by an admin via metastore.php maintenance script.

Bug: T148821
Change-Id: I6c102677ae91715d277e8469597c4f4990b6c928