Only return Status objects from CirrusSearch::searchText
There are times when elasticsearch can return partial results, such as
timeouts. To be able to tell users about these we need to be able to
pass back a Status object that contains both the result set and some
warnings.

Depends on a core patch which changes the assumptions around using
SearchEngine::searchText() to allow for Status objects that are 'OK'
rather than only 'Good'. There are also patches in Flow and
ProofreadPages extensions that will need to be deployed in the same
train.

Change-Id: I34a522bee18d14137d4640b7d5cf2893985b6eea
Depends-On: Ic5e0db727790f4fd189caa54ea5f01672d6a8ea4
Bug: T134157