Use timeouts to limit regex searches rather than max_inspect
max_inspect is a reasonable idea, but unfortunately there is no way to
find out that the search only returned partial results which both
confuses users, and gives them the wrong impression about if/how many
results match their expression.

Remove max_inspect, and instead rely on the combination of the pool
counter and timeouts to limit the impact of regex search on the search
cluster. Timeouts appropriately inform the user that the query could not
be completed, and that they are receiving partial results.

Depends-On: Ic5e0db727790f4fd189caa54ea5f01672d6a8ea4
Bug: T134157
Change-Id: Ibfa6f5f4a411eb8c2f980144e773f8755d05d8bb