Completion preferences radio and content should be side-by-side
When very long text is used the browser ends up putting the radio div
and the content div one above the other. This is a bit of a hackish
solution, but limit the width of the content so that they stay side by
side instead of ending up one above the other.

Bug: T149982
Change-Id: I59fc3f090a63bf41db84e7c6499ceed6a532482e