Add and implement InterwikiResolver
Currently we use only config values to set up the list
of wikis used by crosslanguage and crossproject searches.
This forces to create huge config map to enable to enable
interwiki searches on more wikis.
The solution here is try to infer this data from other components
available in the WMF context.
The idea is to create a new internal service InterwikiResolver
able to provide all information needed for such interwiki searches.
Two implementations are available:

- CirrusConfigInterwikiResolver: can reuse existing config for
  interlanguage and crossproject searches.

- SiteMatrixInterwikiResolver: uses SiteMatix and InterwikiLookup
  to build the map. Added a resources/wmf folder in test/unit
  with all the data needed to run SiteMatrix and InterwikiLookup
  in WMF context.

Third parties can rely on CirrusConfigInterwikiResolver by setting
up the config values and using CirrusConfigInterwikiResolver.

There services are added to MediawikiServices as:
- InterwikiResolverFactory:SERVICE for the factory
- InterwikiResolver::SERVICE for the current wiki config resolver

Copying wmf-config data into test resources is far from ideal but
it was the only way for me to properly develop on top of SiteMatrix.

Bug: T141033
Change-Id: Id2befb1f11809ac9f6af48224695471efd9220a3