Refactor interwiki support
Use the new InterwikiResolver service to handle Interwiki searches.
The goal is to centralize the Elastic Result to Title conversion
into a single place with proper support for interwiki results.
For interwiki results the plan is to rely on the wiki field returned
by elastic to decide if the title is internal or external.
This is still transitional as we need some BC code to support wikis
where the wiki field is not fully populated.
Once done (reindex all wikis) we can remove the BC code marked with a
TODO comment and rely only on the wiki field to determine if the
result refers to an external title.

Added 2 new feature flags to activate either crossproject and
crosslanguage search. These flags are not fully effective, they only
affect behaviors if the config used by CirrusConfigInterwikiResolver
is not present and SiteMatrix is available. This is because it's the
way it works today: if some data are present in some of the map then
the interwiki feature is activated.

The plan is to the SiteMatrix approach for wmf sites and use a
feature flag to activate either crosslanguage fallback and/or
crossproject search.

Bug: T141033
Change-Id: I6aebfa22c05a82f6d887f94ef6c18ce5b3a1b0bc