Properly pass query options when using multi search
This looks like a bug in Elastica where all the search options
(timeout, terminate_after...) are passed in the header section of the
multisearch syntax.  Only few options need to go into the header
others must be passed in the search request body otherwise they are
inneffective.
This patch is backport of a PR suggested to ruflin.

Change-Id: I7a1e818c7ead0188516771f1f4ba82d2937cb8cc