Wait for job to finish processing before doing commons checks
The browser tests have been intermittently failing the commons tests,
and it's quite annoying. The retrys in cindy don't help here, because
the test setup deletes and recreates these pages. If the commons tests
are the only ones being run its probably more likely than in a full test
that the jobs havn't completed.

This is a bit of a hack, to make this better we would probably want to
add an API action that mimics action=cirrusdump, but it will do the
trick for now at least.

Additionally delay the creation of the duplicated file even longer. This
is another sad hack, but if the cirrustest job processes before the
commons job commons never finds out about the duplicated file. It would
be best to have some way to enqueue jobs for another wiki to check this
but not sure it's really possible with mediawiki currently.

Change-Id: I7f3a8e21907ef136038efa507d5150b76e805d1d