Record searchToken in referer for mobile as well
For desktop search we plug a token into the url so it is seen in
referer, and can be used to collect data about what links are clicked
for a particular search. This was only being sent to desktop but in
testing works equally well for the mobile site.

This should allow us to collect more accurate data from which to
measure relevance with the DBN.

Change-Id: I6eeeac426b67d56568bf9686c1e2266e9019de83