Use the core script for running phan
Doesn't change the analysis, just changes where files are located.
It seems keeping things consistent is important. This, combined with
the dependant core patch, allows cirrus to re-use the runner from
core rather than defining it's own script to run phan. This will also
serve as a basis for how other extensions can configure phan to run in
CI.

Don't merge this patch until we have a working CI build that votes
using this. At that point we can merge this patch and turn off Phan
in cindy.

Depends-On: I58ad44f6a8769088e87d5d6cf1079e633c082f88
Bug: T153040
Change-Id: I2671f8cba92d3add0ec1c00c4d2a11516f60de36