Integration of LTR query
This is a first shot at integrating the stored LTR query
implemented in o19s/elasticsearch-learning-to-rank. This adds
a few new search syntax types (which are only sorta-syntax...)
to allow ltr to exclude itself from specific kinds of non-text search
queries.

Adds support for rescore profiles to have fallback based on syntax used
in the query, specifically for the LTR to be able to reject query string
queries. Additionally adds support for fallback chains, rather than a
single level, so we can fallback from LTR to incoming_links +
popularity, to the most basic classic.

Change-Id: I8bb3c67712bcd2601c93a40d2687917d80db5183