Consider full_text a "syntax"
Explicitly add full_text as a syntax and remove special handling of
this query type. This fixes incorrect query type set to "prefix" when
using the prefix keyword in a fulltext search.

Basically the presence of full_text in this array means that the query
displayed search results.

Added a notion of full_text_querystring and full_text_simple_match, if
present.

it means that the fulltext content was inspected, e.g. :
 - full_text,regex: only a regex was present
 - full_text,full_text_querystring,regex: a regex and a fulltext query
   using querystring was used
 - full_text,full_text_simple_match,regex: a regex and a fulltext
   query using the simple match builder was used

This may help to filter queries that we know we can tune, as of today
only queries using the simple_match builder can be tuned.

Interwiki is now using only a cost based filter, if a keyword heavier
than 10 is used then the interwiki search is not performed. This
enables some interesting keyword such as intitle to run with
interwiki.

Sadly I had to keep an heuristic for ResultsType::isSyntaxUsed() which
is used to display the new Title link.

Change-Id: I2f102c3b23e70f81f9fddc1733c27e8bea98a058