Update browsertest search profile for es5
The browser tests were still using a coordination factor. Updating to
es5 removes that coordination factor and caused a few changes to the
browser test results. Update weights so we get the expected results.

This propogates into all the query fixtures, so there are small changes
to all of them.

Bug: T155671
Depends-On: I208a8dbcc2cb05af69af84a8a534ca8e8e03897c
Change-Id: Ia65d11d170c9c782e37944fa84330868c42bea56