Add backend support for the new crossproject result page A/B test
Adds two new config var:
- wgCirrusSearchNewCrossProjectPage: allows to inform SpecialSearch
  that we want to display the new result page
- wgCirrusSearchHideCrossProjectResults: allows to compute recall
  without displaying search results. Will be used in the control
  group to know how many search queries would have displayed
  results for a particular sister wiki

Recall metrics are sent back to the client in the js var
wgCirrusSearchCrossProjectRecall, this is not stictly needed because
we should be able to retrieve this information from hive.
The format is "iwprefix:#res|iwprefix2:#res3". We return the number of
total hits, not the number of results displayed.

Bug: T156299
Change-Id: I3df4f51929ddbfd2bc3544ba6d6e6e2978041a33