Add method to provide custom index settings to IndexCreator
In elasticsearch 5 it is no longer possible to have global configuration
that is applied to all indices (at least, not directly from
elasticsearch.yml). Two options seem to be either provide them at
creation time directly, or create index templates that magically inject
the extra settings. This adds the ability to set custom index settings
in the CirrusSearch config which will be used for all newly created
indices.

This is intentionally against the master branch, rather than es5, so
indices we create between now and then have values explicitly set. We will
still need to run a script to loop over all the existing indices and
set the current slowlog and merge_threads values per-index.

Bug: T155578
Change-Id: I68754ddf1cf7be281981f762cccbdd0f77b9f473