Explicitly use BM25 as default for wmf_defaults similarity profile
The 2.x -> 5.x conversion has some uncertainties aroung similarity
profiles and the 'default' similarity. To make things explicit and
easy to understand force BM25 as the type for wmf_defaults.

Specifically fields in 2.x that specify similarity as default will
report in 5.x that their similarity is classic. It appears though
that the actual search will still use BM25 (tested in 5.1.2, see
https://github.com/elastic/elasticsearch-migration/issues/103#issuecomment-279546227)
To ensure that _mapping doesn't report something wrong, and that
there arn't other lingering, undetected, bugs with this conversion,
we explicitly name BM25 as the similarity config.

Change-Id: I32ddeeb18ac93623ff16b3f9b895c925394e3944