Add an explicit configuration option for nuke script
Last week I accidentaly nuked all the indices on relforge due to
an autogenerated hostname matching some configuration that pointed
it at relforge, when localhost was expected. The nuke script is
automatically run when using update.php, so adjust things such
that it has to be specifically enabled. Vagrant will be updated
to set this for development environments, and labs machines talking
to relforge can have it disabled for safety.

Change-Id: I268a18f644bf4f537f42a08efaf761fab4ec1db3