Fix close redirects, defaultsort and subphrases suggestions
The new completion suggester has a nice feature as it allows
to do retrieve the input value that matched.
It allowed us to remove the double pass mode to fetch redirects.
Unfortunately in some cases we still prefer to force a particular output:
- close redirects that are added to title suggestions, we prefer
  to display the title
- subphrases, those strings are not valid titles
- defaultsort, these are not valid titles

Exception is made of CrossTitleNS where we prefer to display the redirect
encoded as an input rather than the target redirect.

Change-Id: Ic3de0c6d4dbc37f67c158fda4654774143810e33