Add setting for prefix search rescore profile
The switch to bm25 has caused some issues with the prefix search,
specifically sometimes subpages are ranked higher than their parent
page. This looks to be because the weighted sum rescore we switched
to is not well tuned for prefix search. Default prefix search back
to the classic chain which uses multiply rather than weighted sum
to provide a stop-gap solution.

Bug: T156840
Change-Id: I387c60b067d5297f23d3092a54c6240eebf45365