Fix few browser tests
Cindy was running into problems where this waffle squash result was in
the reverse order expected. This test is to make sure that two words
are not squashed together but seems to rely on the phrase rescore.
Remove some wikitext so that the text is no longer in auxiliary_text
and can match phrase rescore.
Concerning filefeature they are sometimes failing because a file
created by another test can be present from time to time.

Change-Id: I1f6c30b40254f8af4f0b871de8aa5f0091e0b185