Allow completion suggester to work with titles that look like integers
Php is quite helpful by doing a wide variety of casting to array keys,
so it's not safe to assume that if you put arbitrary data into an array
key you can take that data back out. In this particular case we were
hitting:

* Strings containing valid integers will be cast to the integer type.

This caused us to pass integers to elasticsearch as the input, and
elasticsearch rejected those inputs and failed to complete building
the completion suggester indices. To fix stop using array keys, they
don't seem to be used for deduplication or anything useful here. Instead
hold onto the text string as part of the data array.

Also fixes an error where the provided $inputDoc is wrong, whatever
doc was last iterated was used as the $inputDoc, instead of the specific
$inputDoc that matches the suggest being generated.

Change-Id: I5b6abe658db0746c763f27b16e4826302899b777