checkIndexes should check index aliased to mw_cirrus_metastore
mw_cirrus_metastore should be an alias, but the script was expecting
it to be an exact alias. Switch things around so the actual alias
is checked.

Also makes a minor adjustment to indentation handling to allow the
format of the new check, no longer resetting $this->path in the
checkIndex() function.

Change-Id: Ic86a4d9de2c70f8721f70384c10b0163d5e26249