Fix problem when using the default similarity
'default' is no longer a similarity provided by elastic
and cannot be used directly.
Cirrus will now throw an error if the similarity for a
specific field cannot be determined.
Fix and deprecated the default profile and set it to classic.
Add a new "classic" profile and make it the default.

Bug: T160301
Change-Id: I3ae80c142f1d14dc54e782319bdf10cd653dc2c1