Temporarily disable smoke:16
Looks like this is the only place we try to show suggestion box, and it doesn't work.
To reduce noise, after manually testing suggestion box is not actually broken, let us
disable it and look for why it may be failing.

The patch also cleans up some duplicate code and weird similar, but not quite identical
scenario strings.

Change-Id: Id1420ae00c5b37b143ce52ee9907ca3392001e0c