Add manual configuration for OtherIndex template boosting
Multimedia search on commonswiki tends to turn up better result than a
multimedia search from other wikis when they query commonswiki. This is
because commons has some template boosts setup the push multimedia that
has been rated by editors as high quality (either via assessments,
featured status, etc).

This adds a hackish feature to allow applying those template boosts to
all sites using commonswiki for multimedia search via the extra indexes
option. This is far from perfect, it won't auto-update when an admin
updates the config on commons, but is relatively easy to do and gives
most of the benefits with minimal work.

There may be a small downside with this, in that wiki-local files will
not be boosted to match the files from commonswiki. This can be worked
around with the 'local:' keyword, but isn't the most ideal. Should be
reasonable though.

Bug: T163223
Change-Id: Ic55dc0cd23c69deda3cacd5f94241db652c4860d