Add rescore fallbacks based on syntax used
The LTR query is going to be slightly more restrictive than existing
rescore profiles as it needs to take in the user search terms and
run queries with it. As such it can't easily support things like phrase
match scores, or the query string syntax. Make it possible to fallback
from ltr to a simpler rescore that supports features the user used.

Change-Id: I65ce25720c83e11438e97dd1da9239bc2591a63c