Don't directly access $GLOBALS in cirrus-config-dump
We are slowly trying to remove any dependence on globals
throughout mediawiki, and instead use the Config objects. It's
probably never going to be done but can't hurt to fix this too.

This is also trivially faster, at 1-2ms vs 7-10ms for execute(), but in
the scheme of things with mediawiki api calls that will probably not
make much difference, as the current api call through the full stack
takes anywhere from 50 to 120ms on terbium.

Change-Id: Id3fa06dccb002981cd0cbb58b432ff2e9c12b786