Fix a series of very minor code smell issues
I found these with the inspection utilities my PHPStorm provides.

The missing "return true" is a little scary. Why doesn't this fail?
It returned "null" and should fail. The reason is because the returned
value is currently entirely unused. But this caller is in core, and
instead of removing the "@return bool" from there, I decided it's
easier to fix this single outlier.

Change-Id: Ia393bb33c2448cf23cd450fec8a080760970f999