Give individual users consistent interleaving
Users generally expect that when they search for something, click
through, and then return to the search page, that results are generally
the same and in the same order. Interleaving breaks that expectation,
as the way the two buckets are interleaved changes on each load.

Fix by seeding the randomness with a token based on the users ip and
user agent. This will give individual users a consistent interleave
while still varying between multiple users.

Change-Id: Ifa51b74481baa169fe7e3d72826a21d399a9c637