Refactor ordering of crosspoject blocks
to allow fine-grained configuration.

Introduce yet another profile system for ordering cross project
blocks. It allows to fine-tune ordering of the blocks based on
different scoring implementations:
- recall
- static
- random

and also permits to combine them using a composite implementation
(weighted sum).
It should then be easier to add other implementations such as a
scoring method based on the density of the highlighted snippets.

Bug: T171803
Change-Id: I9618b9e9bee51d91a9ec931725abcd39eb730b39