Limit the number of tokens in a phrase rescore query
This is a brute force approach to dealing with a problem we have
with the zhwiki analyzer and certain pathological queries. It is
not a complete fix, but will hopefully bandaid over some of its
most problematic effects.

A more complete solution still needs to be determined, perhaps rolling
back or fixing the zh analysis chain.

Bug: T169498
Change-Id: I5d90c40ad5bcf2b648f4299088964c0600e3964d