Add subpageof keyword
It's a specialized version of the prefix keywork to find subpages.
It does not have the limitation of prefix:
- non-greedy
- does not try to handle namespace
- search on redirects

NOTE: The keyword is subpageof, but could be changed to something else.
(feel free to amend the patch).

If reusing the prefix elastic field is not sufficient we'll have to dig
into more dedicated field.

Bug: T159321
Change-Id: Ic1489b00b6179d60d662cd6645d54ba3428343d9