Use correct table to find deletes.
Timestamps in archive table are revision creation timestamp,
and thus unsuitable for retreving deletes by timestamp.

This patch switches to using logging table, where timestamp is the actual
deletion timestamp, plus we check that there is an archive entry for this
title, to ensure the delete wasn't reversed. It is not perfect but it's
reasonable approximation for now.

Bug: T171921
Change-Id: Ie3c3ce292cd6a0c5264157a582544d98b3a6d48b