compsuggest: Handle edge case of empty index
In prod we have a some closed wikis (akwiktionary, probably others)
that have no content pages. As such their titlesuggest index is also
empty. It seems that when the titlesuggest index is empty elasticsearch
short-circuits it's regular response handling and doesn't return the
empty suggest keys that it usually does.

Change-Id: Ifa9ee4eb7f76a1864c5e786c5680d07e487b6e70