Stop eating <span> tags in highlighted output
</span> tags were explicitly excluded from sanitization of highlighting
output, because they are the close marker for highlights. This caused
insource: queries that should return a </span> as part of the highlight
to pass that through to the browser as an html tag, instead of encoded
for display.

This is also a partial implementation of T126563. To complete that
ticket we would want to expose the highlighting tags to the api
endpoints such that non-html users (apps, etc) can correctly handle
highlighting. That is complicated by not wanting to expose this for
cirrussearch without having to implement custom highlight markers for
all the possible search implementations

Bug: T173172
Change-Id: I7a15b2cac65c780dfc03e638734d41c8a4ee1933