Convert integration test api objects to single use
There are some annoying problems with re-using api clients
around api tokens, namely that the tokens can be accidentally
overwritten if multiple things are using it asyncronously.

Convert to a more direct model where new api clients are created
on demand. Also makes it a bit clearer how to do things on multiple
wikis with the api. Multiple wikis on the browser is not handled
here.

Change-Id: I06519bc43f8c17d81920f6210ea2deb1b310d5e4