Remove Messaging Fallback Languages as Language Analysis Fallbacks
Remove code that looks at messaging fallbacks for analysis fallbacks.

Nothing added to list of language/fallback pairs to keep. There was not
a lot of feedback for most, but for the ones that had discussion, it
seemed best not to use the wrong language, even when they are similar.

Update and add additional tests.

Bug: T147959
Change-Id: I87d547d89b6e2159ca9dfa786ca2825abb22dda3