Check multi responses for http errors
It seems 5xx errors don't properly get flagged as such under some
(unknown) circumstances. Double check them so we fail in an expected
way. Prior to this patch the failure would occur by returning a
Status that thought it was OK, but the value was a boolean false
because there are no ResultSet objects in the multi response.

Bug: T167410
Change-Id: I172b6f90e0e35c2f83f8ed809ce4043d76cfc0c3