Improve handling of 5xx responses to elasticsearch requests
The Elastica\Response::hasError() method is almost never the one you
want to check, on its own at least, to see if a request failed. This
only takes into account if the returned json document contains an
'error' key. Other errors, such as a timeout from a proxy in front
of elasticsearch which does not return a json document will not
trigger hasError.

Reviewed current uses of hasError and added checks to Response::isOk()
in appropriate places. Also filed upstream bug at
https://github.com/ruflin/Elastica/issues/1396 to discuss if this
is desired behaviour.

Change-Id: I8677ceb9dcbc7dde455894b60b5e5851f309a74b