Fix create_new_page.feature with ruby
I think this test was broken, the step it refers to does not exist and
some of the queries used could no match to Catapult.
It seems to ignore missing steps.
Flipped amazing-catapult to catapult-amazing because being on a random
we hit the go feature to Amazing Catapult bypassing the srp.

Change-Id: Ibacb14218608e05b9c806171d536581dc89c2a6c