Port create_new_page.feature to nodejs
Silenced logs in world.js
Not sure that how I export page class is right, tbh I don't really
understand. We export the class when we want to inherit but an
instance when we want to use methods statically?
Not entirely sure either if the way to test the page in the browser is
right... well it seems to work.

Not a big fan of chai to.be.true. Spent some time figuring out why it
did not work and was because I wrote is.true that was silently
ignored. I replaced with to.equal(true|false|undefined) that seems more
concise since we save the jshints.

Change-Id: I1fb6c5fa294d7480841502277fc508397fe50c92